cook

+++ src/liangGraphs/Displayable.java
@@ -1,7 +1,7 @@
 package liangGraphs;
 
 public interface Displayable {
-    public double getX(); // Get x-coordinate of the vertex
+    public int getX(); // Get x-coordinate of the vertex
 
 public int getY(); // Get y-coordinate of the vertex
 public String getName();
